Complete venv removal
Removal of the venv has made bootstrapping and
installing bifrost significantly easier. Instead
of using a venv and polluting the system with pip
based installation, we're using Ansible installed
via epel-release.

Deployment of bifrost is much more consistent, and
the import of the hardware into ironic is working
exactly as it should.